+++ CHANGELOG.md
@@ -32,6 +32,7 @@ All notable changes to this project will be documented in this file.
 - Commandline replacing sound stream block ranges
 - [#1625] Error log frame - Save all to file button
 - [#2467] Show ImportAssets name/characterId in the title when there is only single one item
+- [#2468] Option for `-dumpas2` CLI command to use export names (`-exportNames` option)
 
 ### Changed
 - AS1/2 - Single DoAction tag inside frame is now displayed directly as frame node
@@ -3861,6 +3862,7 @@ Major version of SWF to XML export changed to 2.
 [#2451]: https://www.free-decompiler.com/flash/issues/2451
 [#1625]: https://www.free-decompiler.com/flash/issues/1625
 [#2467]: https://www.free-decompiler.com/flash/issues/2467
+[#2468]: https://www.free-decompiler.com/flash/issues/2468
 [#2456]: https://www.free-decompiler.com/flash/issues/2456
 [#2459]: https://www.free-decompiler.com/flash/issues/2459
 [#2460]: https://www.free-decompiler.com/flash/issues/2460